refactor(database): share PostgreSQL infrastructure
