+++ agent/analytics/capture.py
@@ -4,7 +4,7 @@ import logging
 from collections.abc import Awaitable, Callable
 from functools import wraps
 
-from agent.analytics.database import configured
+from agent.database.analytics import configured
 
 logger = logging.getLogger(__name__)
