+++ agent/analytics/identity.py
@@ -2,8 +2,8 @@
 
 from uuid import UUID
 
-from agent.analytics.database import configured, workspace_id
 from agent.analytics.events import person_uuid, subject_uuid
+from agent.database.analytics import configured, workspace_id
 
 
 def opaque_id(kind: str, value: str | int | None) -> UUID | None:
